code/fastapi/requirements.txt
fastapi>=0.95.0,<1.0.0
uvicorn>=0.20.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0,<2.0.0
pydantic>=1.10.0,<2.0.0